Improve line bubble rendering
The text blob bounds that are provided were not accurate to the size of the
text. I discovered another method which provides more accurate text measurements
that is being used insted. This change also enables anti-aliasing for the line
bubbles. Both of these changes together improves the appearance of line bubbles
significantly.